Crud completo usuario/rol, se agrego dependencia spring security

+++ src/main/java/com/stockweb/demo/core/repository/RolRepository.java
@@ -3,6 +3,5 @@ package com.stockweb.demo.core.repository;
 import com.stockweb.demo.core.model.Rol;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolRepository extends JpaRepository<Rol,Integer> {
-
+public interface RolRepository extends JpaRepository<Rol,Long> {
 }
